Convert spec URLs test to linter unit
